Stream scenes incrementally, crossfade skies, and keep physics per frame
Three pieces of work that share files, committed together.

Incremental scenes. `Scene::flattenHierarchy`, which rebuilt every render,
simulation and resource list on any hierarchy change anywhere, is replaced by a
per-scene `SceneIndex`: nodes carry subtree, local, resource and transform
revisions that propagate along the ancestor path only, unchanged branches are
skipped, and resource ownership is counted incrementally and published to the
GPU budget when it changes. Iteration order of the dense lists is unspecified.
Alongside it: `Node::setVisible` (render-only visibility that keeps simulation
and resources), a configurable `GeometryCapacity` at Engine/ResourceManager
construction with unchanged defaults, `Scene::rebaseOrigin`/`rebaseSubtree`
for floating-origin worlds (bodies, characters and joint anchors follow), LOD
Group selecting child representations by projected coverage with hysteresis
(shared with MeshNode's chain), and `Behaviour::onPhysicsStep` so the vehicle
applies its forces on each fixed 1/60 s step instead of the frame delta.

Skies that change with the hour. The skybox crossfades two skies
(`skyboxBlendTexture`, `skyboxBlend`, `skyboxBlendRotation`, each sky with its
own rotation) and can draw a Sun disc over them (`skySunDirection`,
`skySunColor`, `skySunSize`) for photographed skies whose own Sun was removed.
Scripts swap skies by path with `scene.setSkybox`, which refuses a path outside
the project or with no file behind it and logs why. The sky pass clamps its
output under the half-float ceiling of the HDR target so it cannot store
infinity. Saved by path, snapshotted, counted as resources, in the Inspector;
the fold-determinism reference gains the new keys.

Physics that costs what moves. Node transforms are synced once per frame after
the last substep, not after each one (the vehicle reads its pose from Jolt);
only behaviours declaring `hasPhysicsStep` and bodies declaring
`hasPrePhysicsStep` are called per substep; dynamic bodies are synced back only
if Jolt had them awake during the frame; static and kinematic bodies are pushed
to Jolt only when their world transform changed, which spares a broadphase
update per static collider per frame; body access uses Jolt's lock-free
interfaces, sound because bodies are only touched from the main thread between
updates.

+++ src/render/Renderer.hpp
@@ -260,8 +260,8 @@ private:
     int giBakeFramesRemaining_ = 0;
     int giRealtimeWarmupRemaining_ = kGIRealtimeWarmupFrames;
     uint64_t giFrameCounter_ = 0;
-    uint32_t giLastHierarchyVersion_ = 0;
-    uint32_t giLastTransformVersion_ = 0;
+    uint64_t giLastHierarchyVersion_ = 0;
+    uint64_t giLastTransformVersion_ = 0;
     uint64_t giLastDirtySignature_ = 0;
     bool giWasEnabled_ = true;
     int giLastLightingMode_ = 0;
